Polish
To-do:
Follow AI.
1. Instance disappears when their maximum range is reached.
2. Timer starts when instance reaches maximum range, when timer ends
instance dissappears.
3. Radius determined range.
Bugs:
For some reason the playable character has to be put halfway into the
ground, and the capsule collider has to be raised as well, for the
character to be placed correctly when the game starts.
Animations are a bit jerky, which is mostly likely just a matter of
adjusting start/end times and looping.
Backwards walk is especially strange though, one of the feet gets stuck.