Fixed head turning bug
Fixed a head turning bug, which was caused by the head angle lerp after
the player changed orientation.
E.g.: If the player turns 180 degrees, the head will slowly lerp from
where it was facing to the new orientation.
This has been fixed by checking for the delta angle, and if it exceeds
80 degrees, the lerp is replaced by a simple assignment to the
destination rotation.