Added aiming & shooting
Added a proof of concept aiming & shooting system.
A ray is sent out from the middle of the screen into the z-direction.
The point where it hits is recorded, and used to create a direction
vector from the middle of the player to the hit point.
This direction vector is then used to set the velocity of a projectile's
rigid body.
The Player.prefab is changed to include the new script.