Level Art Finally uploaded
This is the repo for level art and co. It now contains the general
placements of thing for the 2 first part of the level 01. PLEASE NOTICE
that none of the object present are the final object...they are just
stand in. I have notice some minor issues about the scalling of
house_02, TrainContainer....etc. If you see anything that looks wack
please give a shout !!!